Split into separate classes and packages
