+++ experiments/ambleck/src/main/java/uk/co/metagrid/ambleck/webapp/AmbleckApplication.java
@@ -1,0 +1,47 @@
+/*
+ *
+ * <meta:header>
+ *   <meta:licence>
+ *     Copyright (C) 2024 University of Manchester.
+ *
+ *     This information is free software: you can redistribute it and/or modify
+ *     it under the terms of the GNU General Public License as published by
+ *     the Free Software Foundation, either version 3 of the License, or
+ *     (at your option) any later version.
+ *
+ *     This information is distributed in the hope that it will be useful,
+ *     but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *     GNU General Public License for more details.
+ *
+ *     You should have received a copy of the GNU General Public License
+ *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *   </meta:licence>
+ * </meta:header>
+ *
+ * Using the @SpringBootApplication Annotation
+ * https://docs.spring.io/spring-boot/docs/2.0.x/reference/html/using-boot-using-springbootapplication-annotation.html
+ *
+ */
+
+package uk.co.metagrid.ambleck.webapp;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
+import uk.co.metagrid.ambleck.util.YamlConverter;
+
+@SpringBootApplication
+@Import(
+    { YamlConverter.class }
+    )
+public class AmbleckApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(
+		    AmbleckApplication.class,
+		    args
+		    );
+	    }
+    }
+
