+++ experiments/ambleck/src/main/java/uk/co/metagrid/ambleck/webapp/ServletInitializer.java
@@ -21,19 +21,17 @@
  *
  */
 
-package uk.co.metagrid.ambleck;
+package uk.co.metagrid.ambleck.webapp;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
 
-@SpringBootApplication
-public class AmbleckApplication {
+public class ServletInitializer extends SpringBootServletInitializer {
 
-	public static void main(String[] args) {
-		SpringApplication.run(
-		    AmbleckApplication.class,
-		    args
+	@Override
+	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+		return application.sources(
+		    AmbleckApplication.class
 		    );
 	    }
     }
-
